secp256k1: Move sig benchmarks to separate file.
